requirements.txt
torch>=1.11.0
torch-geometric>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
networkx>=2.6.0
plotly>=5.0.0
tqdm>=4.62.0